refactor(jwt): move signing/JWKS files into signing/ and refresh token into refresh/
Restructure packages/jwt/src/ by moving signer, verifier, and JWKS
files into a signing/ subdirectory and refresh token files into a
refresh/ subdirectory. All internal import paths are updated; the
public API (index.ts exports) remains identical.

Closes #856
