+++ packages/jwt/src/index.ts
@@ -1,9 +1,9 @@
 export * from './errors.js';
-export * from './jwks.js';
+export * from './signing/jwks.js';
 export * from './module.js';
-export * from './refresh-token.js';
+export * from './refresh/refresh-token.js';
 export * from './service.js';
-export * from './signer.js';
+export * from './signing/signer.js';
 export * from './status.js';
 export * from './types.js';
-export * from './verifier.js';
+export * from './signing/verifier.js';
